Revamp skybox pass
