+++ rend3-routine/src/common/mod.rs
@@ -3,5 +3,4 @@ pub mod forward_pass;
 pub mod interfaces;
 pub mod samplers;
 pub mod shaders;
-pub mod skybox_pass;
 pub mod sorting;
