+++ rend3-routine/src/lib.rs
@@ -597,193 +597,7 @@ impl PrimaryPasses {
     }
 }
 
-pub struct SkyboxRoutine {
-    pub interfaces: common::interfaces::ShaderInterfaces,
-    pub skybox_pass: skybox::SkyboxPass,
-    pub options: RenderTextureOptions,
 
-    pub skybox_texture: Option<TextureHandle>,
-}
-
-impl SkyboxRoutine {
-    pub fn new(renderer: &Renderer, options: RenderTextureOptions) -> Self {
-        // TODO: clean up
-        let interfaces = common::interfaces::ShaderInterfaces::new(&renderer.device, renderer.mode);
-
-        let skybox_pipeline = common::skybox_pass::build_skybox_pipeline(common::skybox_pass::BuildSkyboxShaderArgs {
-            mode: renderer.mode,
-            device: &renderer.device,
-            interfaces: &interfaces,
-            samples: options.samples,
-        });
-
-        let skybox_pass = skybox::SkyboxPass::new(skybox_pipeline);
-
-        Self {
-            skybox_pass,
-            options,
-            interfaces,
-            skybox_texture: None,
-        }
-    }
-
-    pub fn set_background_texture(&mut self, texture: Option<TextureHandle>) {
-        self.skybox_texture = texture;
-    }
-
-    pub fn resize(&mut self, renderer: &Renderer, options: RenderTextureOptions) {
-        if self.options.samples != options.samples {
-            let skybox_pipeline =
-                common::skybox_pass::build_skybox_pipeline(common::skybox_pass::BuildSkyboxShaderArgs {
-                    mode: renderer.mode,
-                    device: &renderer.device,
-                    interfaces: &self.interfaces,
-                    samples: options.samples,
-                });
-            self.skybox_pass = skybox::SkyboxPass::new(skybox_pipeline);
-        }
-
-        self.options = options;
-    }
-
-    pub fn ready(&mut self, renderer: &Renderer) {
-        let d2c_texture_manager = renderer.d2c_texture_manager.read();
-
-        profiling::scope!("Update Skybox");
-
-        if let Some(handle) = self.skybox_texture {
-            if self.skybox_pass.current_skybox.as_ref().map(|s| &s.handle) == Some(&handle) {
-                return;
-            }
-
-            let bg = BindGroupBuilder::new()
-                .append_texture_view(d2c_texture_manager.get_view(handle.get_raw()))
-                .build(&renderer.device, Some("skybox"), &self.interfaces.skybox_bgl);
-
-            self.skybox_pass.current_skybox = Some(StoredSkybox { bg, handle })
-        }
-    }
-
-    pub fn add_to_graph<'node>(
-        &'node self,
-        graph: &mut RenderGraph<'node>,
-        color: RenderTargetHandle,
-        resolve: Option<RenderTargetHandle>,
-        depth: RenderTargetHandle,
-        forward_uniform_bg: DataHandle<BindGroup>,
-    ) {
-        let mut builder = graph.add_node("Skybox");
-
-        let hdr_color_handle = builder.add_render_target_output(color);
-        let hdr_resolve = builder.add_optional_render_target_output(resolve);
-        let hdr_depth_handle = builder.add_render_target_input(depth);
-
-        let rpass_handle = builder.add_renderpass(RenderPassTargets {
-            targets: vec![RenderPassTarget {
-                color: hdr_color_handle,
-                clear: Color::BLACK,
-                resolve: hdr_resolve,
-            }],
-            depth_stencil: Some(RenderPassDepthTarget {
-                target: DepthHandle::RenderTarget(hdr_depth_handle),
-                depth_clear: Some(0.0),
-                stencil_clear: None,
-            }),
-        });
-
-        let forward_uniform_handle = builder.add_data_input(forward_uniform_bg);
-        let pt_handle = builder.passthrough_ref(self);
-
-        builder.build(move |pt, _renderer, encoder_or_pass, temps, _ready, graph_data| {
-            let this = pt.get(pt_handle);
-            let rpass = encoder_or_pass.get_rpass(rpass_handle);
-
-            let forward_uniform_bg = graph_data.get_data(temps, forward_uniform_handle).unwrap();
-
-            if let Some(ref skybox) = this.skybox_pass.current_skybox {
-                rpass.set_pipeline(&this.skybox_pass.skybox_pipeline);
-                rpass.set_bind_group(0, forward_uniform_bg, &[]);
-                rpass.set_bind_group(1, &skybox.bg, &[]);
-                rpass.draw(0..3, 0..1);
-            }
-        });
-    }
-}
-
-pub struct TonemappingRoutine {
-    pub interfaces: common::interfaces::ShaderInterfaces,
-    pub tonemapping_pass: tonemapping::TonemappingPass,
-    pub size: UVec2,
-}
-
-impl TonemappingRoutine {
-    pub fn new(renderer: &Renderer, size: UVec2, output_format: TextureFormat) -> Self {
-        // TODO: clean up
-        let interfaces = common::interfaces::ShaderInterfaces::new(&renderer.device, renderer.mode);
-
-        let tonemapping_pass = tonemapping::TonemappingPass::new(tonemapping::TonemappingPassNewArgs {
-            device: &renderer.device,
-            interfaces: &interfaces,
-            output_format,
-        });
-
-        Self {
-            tonemapping_pass,
-            size,
-            interfaces,
-        }
-    }
-
-    pub fn resize(&mut self, size: UVec2) {
-        self.size = size;
-    }
-
-    pub fn add_to_graph<'node>(
-        &'node self,
-        graph: &mut RenderGraph<'node>,
-        src: RenderTargetHandle,
-        dst: RenderTargetHandle,
-        forward_uniform_bg: DataHandle<BindGroup>,
-    ) {
-        let mut builder = graph.add_node("Tonemapping");
-
-        let input_handle = builder.add_render_target_input(src);
-        let output_handle = builder.add_render_target_output(dst);
-
-        let rpass_handle = builder.add_renderpass(RenderPassTargets {
-            targets: vec![RenderPassTarget {
-                color: output_handle,
-                clear: Color::BLACK,
-                resolve: None,
-            }],
-            depth_stencil: None,
-        });
-
-        let forward_uniform_handle = builder.add_data_input(forward_uniform_bg);
-
-        let pt_handle = builder.passthrough_ref(self);
-
-        builder.build(move |pt, renderer, encoder_or_pass, temps, _ready, graph_data| {
-            let this = pt.get(pt_handle);
-            let rpass = encoder_or_pass.get_rpass(rpass_handle);
-            let forward_uniform_bg = graph_data.get_data(temps, forward_uniform_handle).unwrap();
-            let hdr_color = graph_data.get_render_target(input_handle);
-
-            profiling::scope!("tonemapping");
-
-            let blit_src_bg = temps.add(BindGroupBuilder::new().append_texture_view(hdr_color).build(
-                &renderer.device,
-                Some("blit src bg"),
-                &this.interfaces.blit_bgl,
-            ));
-
-            rpass.set_pipeline(&this.tonemapping_pass.pipeline);
-            rpass.set_bind_group(0, forward_uniform_bg, &[]);
-            rpass.set_bind_group(1, blit_src_bg, &[]);
-            rpass.draw(0..3, 0..1);
-        });
-    }
-}
 
 struct PerTransparencyInfo {
     ty: TransparencyType,
